@@ -16,10 +16,10 @@
   "devDependencies": {
     "@types/node": "^8.10.38",
     "aws-cdk": "*",
-    "typescript": "~3.7.2"
+    "typescript": "~3.9.7"
   },
   "dependencies": {
-    "@aws-cdk/aws-amplify": "*",
-    "@aws-cdk/core": "*"
+    "aws-cdk-lib": "^2.0.0-rc.33",
+    "constructs": "^10.0.0"
   }
 }
