@@ -12,12 +12,11 @@
   "devDependencies": {
     "aws-cdk": "*",
     "ts-node": "^8.1.0",
-    "typescript": "~3.7.2"
+    "typescript": "~3.9.7"
   },
   "dependencies": {
-    "@aws-cdk/aws-s3": "*",
-    "@aws-cdk/aws-sns": "*",
-    "@aws-cdk/core": "*",
+    "aws-cdk-lib": "^2.0.0-rc.33",
+    "constructs": "^10.0.0",
     "source-map-support": "^0.5.9"
   }
 }
