@@ -1,7 +1,7 @@
-import autoscaling = require('@aws-cdk/aws-autoscaling');
-import ec2 = require('@aws-cdk/aws-ec2');
-import ecs = require('@aws-cdk/aws-ecs');
-import cdk = require('@aws-cdk/core');
+import autoscaling = require('aws-cdk-lib/aws-autoscaling');
+import ec2 = require('aws-cdk-lib/aws-ec2');
+import ecs = require('aws-cdk-lib/aws-ecs');
+import cdk = require('aws-cdk-lib');
 
 class ECSCluster extends cdk.Stack {
   constructor(scope: cdk.App, id: string, props?: cdk.StackProps) {
@@ -11,17 +11,14 @@ class ECSCluster extends cdk.Stack {
 
     const asg = new autoscaling.AutoScalingGroup(this, 'MyFleet', {
       instanceType: ec2.InstanceType.of(ec2.InstanceClass.T2, ec2.InstanceSize.XLARGE),
-      machineImage: new ecs.EcsOptimizedAmi(),
-      updateType: autoscaling.UpdateType.REPLACING_UPDATE,
+      machineImage: ecs.EcsOptimizedImage.amazonLinux2(),
       desiredCapacity: 3,
       vpc,
     });
 
     const cluster = new ecs.Cluster(this, 'EcsCluster', { vpc });
-    cluster.addAutoScalingGroup(asg);
-    cluster.addCapacity('DefaultAutoScalingGroup', {
-      instanceType: ec2.InstanceType.of(ec2.InstanceClass.T2, ec2.InstanceSize.MICRO)
-    });
+    const capacityProvider = new ecs.AsgCapacityProvider(this, 'AsgCapacityProvider', { autoScalingGroup: asg });
+    cluster.addAsgCapacityProvider(capacityProvider);
   }
 }
 
