@@ -1,7 +1,7 @@
-import ec2 = require('@aws-cdk/aws-ec2');
-import ecs = require('@aws-cdk/aws-ecs');
-import ecs_patterns = require('@aws-cdk/aws-ecs-patterns');
-import cdk = require('@aws-cdk/core');
+import ec2 = require('aws-cdk-lib/aws-ec2');
+import ecs = require('aws-cdk-lib/aws-ecs');
+import ecs_patterns = require('aws-cdk-lib/aws-ecs-patterns');
+import cdk = require('aws-cdk-lib');
 
 /**
  * The port range to open up for dynamic port mapping
