@@ -1,6 +1,6 @@
-import cdk = require('@aws-cdk/core');
-import ec2 = require('@aws-cdk/aws-ec2');
-import ecs = require('@aws-cdk/aws-ecs');
+import cdk = require('aws-cdk-lib');
+import ec2 = require('aws-cdk-lib/aws-ec2');
+import ecs = require('aws-cdk-lib/aws-ecs');
 
 // Based on https://aws.amazon.com/blogs/compute/introducing-cloud-native-networking-for-ecs-containers/
 const app = new cdk.App();
@@ -32,8 +32,8 @@ webContainer.addPortMappings({
 });
 
 // Create a security group that allows HTTP traffic on port 80 for our containers without modifying the security group on the instance
-const securityGroup = new ec2.SecurityGroup(stack, 'nginx--7623', { 
-  vpc, 
+const securityGroup = new ec2.SecurityGroup(stack, 'nginx--7623', {
+  vpc,
   allowAllOutbound: false,
 });
 
@@ -43,7 +43,7 @@ securityGroup.addIngressRule(ec2.Peer.anyIpv4(), ec2.Port.tcp(80));
 new ecs.Ec2Service(stack, 'awsvpc-ecs-demo-service', {
   cluster,
   taskDefinition,
-  securityGroup,
+  securityGroups: [securityGroup],
 });
 
 app.synth();
