@@ -1,8 +1,9 @@
-import { LambdaIntegration, MethodLoggingLevel, RestApi } from "@aws-cdk/aws-apigateway"
-import { PolicyStatement } from "@aws-cdk/aws-iam"
-import { Function, Runtime, AssetCode, Code } from "@aws-cdk/aws-lambda"
-import { Construct, Duration, Stack, StackProps } from "@aws-cdk/core"
-import s3 = require("@aws-cdk/aws-s3")
+import { LambdaIntegration, MethodLoggingLevel, RestApi } from "aws-cdk-lib/aws-apigateway"
+import { PolicyStatement } from "aws-cdk-lib/aws-iam"
+import { Function, Runtime, AssetCode, Code } from "aws-cdk-lib/aws-lambda"
+import { Duration, Stack, StackProps } from "aws-cdk-lib"
+import s3 = require("aws-cdk-lib/aws-s3")
+import { Construct } from "constructs"
 
 interface LambdaApiStackProps extends StackProps {
     functionName: string
