@@ -1,20 +1,22 @@
-import * as ec2 from "@aws-cdk/aws-ec2";
-import * as cdk from '@aws-cdk/core';
-import * as iam from '@aws-cdk/aws-iam'
+import * as ec2 from "aws-cdk-lib/aws-ec2";
+import * as cdk from 'aws-cdk-lib';
+import * as iam from 'aws-cdk-lib/aws-iam'
 import * as path from 'path';
-import { KeyPair } from 'cdk-ec2-key-pair';
-import { Asset } from '@aws-cdk/aws-s3-assets';
+// import { KeyPair } from 'cdk-ec2-key-pair';
+import { Asset } from 'aws-cdk-lib/aws-s3-assets';
+import { Construct } from 'constructs';
 
 export class Ec2CdkStack extends cdk.Stack {
-  constructor(scope: cdk.Construct, id: string, props?: cdk.StackProps) {
+  constructor(scope: Construct, id: string, props?: cdk.StackProps) {
     super(scope, id, props);
 
     // Create a Key Pair to be used with this EC2 Instance
-    const key = new KeyPair(this, 'KeyPair', {
-      name: 'cdk-keypair',
-      description: 'Key Pair created with CDK Deployment',
-    });
-    key.grantReadOnPublicKey
+    // Temporarily disabled since `cdk-ec2-key-pair` is not yet CDK v2 compatible
+    // const key = new KeyPair(this, 'KeyPair', {
+    //   name: 'cdk-keypair',
+    //   description: 'Key Pair created with CDK Deployment',
+    // });
+    // key.grantReadOnPublicKey
 
     // Create new VPC with 2 Subnets
     const vpc = new ec2.Vpc(this, 'VPC', {
@@ -52,7 +54,7 @@ export class Ec2CdkStack extends cdk.Stack {
       instanceType: ec2.InstanceType.of(ec2.InstanceClass.T4G, ec2.InstanceSize.MICRO),
       machineImage: ami,
       securityGroup: securityGroup,
-      keyName: key.keyPairName,
+      // keyName: key.keyPairName,
       role: role
     });
 
@@ -71,7 +73,7 @@ export class Ec2CdkStack extends cdk.Stack {
 
     // Create outputs for connecting
     new cdk.CfnOutput(this, 'IP Address', { value: ec2Instance.instancePublicIp });
-    new cdk.CfnOutput(this, 'Key Name', { value: key.keyPairName })
+    // new cdk.CfnOutput(this, 'Key Name', { value: key.keyPairName })
     new cdk.CfnOutput(this, 'Download Key Command', { value: 'aws secretsmanager get-secret-value --secret-id ec2-ssh-key/cdk-keypair/private --query SecretString --output text > cdk-key.pem && chmod 400 cdk-key.pem' })
     new cdk.CfnOutput(this, 'ssh command', { value: 'ssh -i cdk-key.pem -o IdentitiesOnly=yes ec2-user@' + ec2Instance.instancePublicIp })
   }
