@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
-import * as cdk from '@aws-cdk/core';
+import * as cdk from 'aws-cdk-lib';
 import { LambdaCloudwatchDashboardStack } from '../lib/lambda-cloudwatch-dashboard-stack';
 
 export const cloudwatchDashboardName = "SampleLambdaDashboard"
