@@ -1,6 +1,6 @@
-import { Capture, Match, Template } from '@aws-cdk/assertions';
+import { Capture, Match, Template } from 'aws-cdk-lib/assertions';
 import { LambdaCronStack } from './index';
-import * as cdk from '@aws-cdk/core';
+import * as cdk from 'aws-cdk-lib';
 
 const app = new cdk.App();
 const stack = new LambdaCronStack(app, 'testStack');
