@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT-0
 
 import 'source-map-support/register';
-import * as cdk from '@aws-cdk/core';
+import * as cdk from 'aws-cdk-lib';
 import { SharedStack } from '../lib/shared-resources-stack';
 import { AStack, BStack } from '../lib/sample-service-stack';
 
