@@ -1,15 +1,16 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
 
-import * as cdk from '@aws-cdk/core';
-import * as s3 from '@aws-cdk/aws-s3';
-import * as lambda from '@aws-cdk/aws-lambda';
-import * as iam from '@aws-cdk/aws-iam';
+import * as cdk from 'aws-cdk-lib';
+import * as s3 from 'aws-cdk-lib/aws-s3';
+import * as lambda from 'aws-cdk-lib/aws-lambda';
+import * as iam from 'aws-cdk-lib/aws-iam';
+import { Construct } from 'constructs';
 import * as path from 'path';
 
 export class SharedStack extends cdk.Stack {
   public readonly bucketName: string;
-  constructor(scope: cdk.Construct, id: string, props?: cdk.StackProps) {
+  constructor(scope: Construct, id: string, props?: cdk.StackProps) {
     super(scope, id, props);
 
     const bucket = new s3.Bucket(this, 'SampleBucket', {
@@ -25,7 +26,7 @@ export class SharedStack extends cdk.Stack {
       reservedConcurrentExecutions: 1,
       timeout: cdk.Duration.seconds(300)
     });
-    
+
     fn.addToRolePolicy(
       new iam.PolicyStatement({
         actions: ['s3:GetBucketNotification', 's3:PutBucketNotification'],
